Mostly changes to grass field agent.

+++ src/chalmers/dax021308/ecosystem/model/agent/AbstractAgent.java
@@ -32,7 +32,7 @@ public abstract class AbstractAgent implements IAgent {
 	protected int height;
 	protected int capacity = Integer.MAX_VALUE;
 	protected int lifeLength = 0;
-	protected int energy = 1000; // set specific energy level in subclasses
+	protected double energy = 1000; // set specific energy level in subclasses
 	protected int trophicLevel;
 	protected Vector velocity;
 	protected Gender gender;
